+++ src/main/java/com/hbm/render/tileentity/RenderZirnox.java
@@ -17,13 +17,13 @@ public class RenderZirnox extends TileEntitySpecialRenderer {
 		GL11.glDisable(GL11.GL_CULL_FACE);
 		switch(tileEntity.getBlockMetadata() - 10) {
 		case 2:
-			GL11.glRotatef(0, 0F, 1F, 0F); break;
+			GL11.glRotatef(270, 0F, 1F, 0F); break;
 		case 3:
 			GL11.glRotatef(90, 0F, 1F, 0F); break;
 		case 4:
 			GL11.glRotatef(180, 0F, 1F, 0F); break;
 		case 5:
-			GL11.glRotatef(270, 0F, 1F, 0F); break;
+			GL11.glRotatef(0, 0F, 1F, 0F); break;
 		}
 		
 		GL11.glShadeModel(GL11.GL_SMOOTH);
